VQA/backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-multipart==0.0.20
python-dotenv==1.1.0
google-generativeai==0.8.5
pillow==11.2.1
streamlit==1.46.0